moved naming logic to another pkg
